fix(a2a): coworker-tasking is visibility-only, remove human summon-entry UX (EP-A2A)
G2 was implemented as a human entry form (CoworkerSummonPicker: coworker
dropdown + "what should they do?" field), putting the burden of choosing and
tasking peers on the human. Per founder correction (2026-06-06), that is the
active coworker's responsibility, not the human's. The portal should only SHOW
what the active coworker is tasking and a summary of what each peer is doing.

- Delete CoworkerSummonPicker and the user-facing summonCoworkerAction /
  listSummonableCoworkers (also an auth()-only governance side door); replace
  with read-only conversation-participants-action.ts.
- AgentCoworkerPanel renders only the read-only CollaborationActivityPanel;
  summon/handoff cards attribute the active coworker as source, not "You".
- summon_coworker is now explicitly coworker-initiated, carrying callerAgentId;
  core + event bus thread the attribution through.
- Spec + Slice 1 plan reframed G2 to visibility-only; reconciled built table,
  acceptance, risks, and open questions.

Underlying request_coworker / summon_coworker mechanism unchanged.
